Remove build warnings and vendor TF Lite API AAR

+++ braillekeyboard/src/main/cpp/liblouis/liblouis/metadata.c
@@ -645,7 +645,7 @@ parseQuery(const char *query) {
 	if (!unicodeRange) {
 		// default value of unicode-range is determined by CHARSIZE
 		static char value[5] = "";
-		if (!*value) sprintf(value, "ucs%ld", CHARSIZE);
+                if (!*value) sprintf(value, "ucs%u", (unsigned int)CHARSIZE);
 		FeatureWithImportance *f = memcpy(malloc(sizeof(FeatureWithImportance)),
 				(&(FeatureWithImportance){
 						feat_new(strdup("unicode-range"), strdup(value),
